Retro replay (#111)
This PR implements retroReplay -- replay an execution and everything after it, using the original transaction log for the execution order. In order to implement this feature, I made several changes:

- Changed isReplay to replayMode. 0: not a replay, 1: replay a single execution using the provenance log, 2: replay an execution and everything after.
- Added an option to log requests. We only need to log the initial client input of each workflow. It uses the provenance buffer so it should not impact performance.
- The Apiary worker uses a single thread to do retroReplay and returns the result once everything is finished. -RetroReplay does not generate more TCP requests to the worker.
- Added tests to make sure that we can faithfully follow the original execution order, especially if two requests are interleaving.

Now we can reset the database and re-execute from any point in the past. Will add more optimizations on selective replay in future PRs.