[Debugger] WordPress example (#115)
This PR implements WordPress-11073, where the bug fix actually introduced a race condition:

While trashing a post and its comments, it contains two steps: first back up comments in a postmeta table and then mark all comments as trashed. if a new comment arrives in between these two steps, the new comment would not be recorded in postmeta table but would be marked as trashed in the second step.
Then if we need to untrash the post and restore its comments, the code will look up postmeta and try to unmark those trashed posts. However, the latest comment will not show up because it is still marked as trashed.

In order to simulate interleaving executions of functions, this PR introduced workerAsyncDelay in ApiaryConifg. So it will be easier to reproduce race conditions due to non-transactional workflow executions.

This PR also fixed an error handling bug in PostgresConnection.java: the code did not capture or retry serialization failure if the exception occurred during the commit phase. So we have to check if the outermost level exception is PSQLException as well.

One issue I also realized is that serializable isolation in Postgres cannot guarantee that the transaction ID order is the same as the actual transaction schedule order! For example, transaction A has a smaller transaction ID than B and A started first. However, the actual schedule could be B -> A. So even if the transaction log shows interleaved transactions, the actual order might not be interleaved at all. This may be just a Postgres issue but could lead to some unexpected outcomes.

The transaction order issue may also cause random test failures. Will fix this in the future PR.